crc_fast

Enum CrcAlgorithm

pub enum CrcAlgorithm {

    CrcCustom,
    Crc16Arc,
    Crc16Cdma2000,
    Crc16Cms,
    Crc16Dds110,
    Crc16DectR,
    Crc16DectX,
    Crc16Dnp,
    Crc16En13757,
    Crc16Genibus,
    Crc16Gsm,
    Crc16Ibm3740,
    Crc16IbmSdlc,
    Crc16IsoIec144433A,
    Crc16Kermit,
    Crc16Lj1200,
    Crc16M17,
    Crc16MaximDow,
    Crc16Mcrf4xx,
    Crc16Modbus,
    Crc16Nrsc5,
    Crc16OpensafetyA,
    Crc16OpensafetyB,
    Crc16Profibus,
    Crc16Riello,
    Crc16SpiFujitsu,
    Crc16T10Dif,
    Crc16Teledisk,
    Crc16Tms37157,
    Crc16Umts,
    Crc16Usb,
    Crc16Xmodem,
    Crc32Aixm,
    Crc32Autosar,
    Crc32Base91D,
    Crc32Bzip2,
    Crc32CdRomEdc,
    Crc32Cksum,
    Crc32Custom,
    Crc32Iscsi,
    Crc32IsoHdlc,
    Crc32Jamcrc,
    Crc32Mef,
    Crc32Mpeg2,
    Crc32Xfer,
    Crc64Custom,
    Crc64Ecma182,
    Crc64GoIso,
    Crc64Ms,
    Crc64Nvme,
    Crc64Redis,
    Crc64We,
    Crc64Xz,

}

Supported CRC-16, CRC-32, and CRC-64 variants

Variants

CrcCustom

Generic custom CRC variant that works with any supported width (16, 32, 64). The actual width is determined by the width field in CrcParams.

Crc16Arc

Crc16Cdma2000

Crc16Cms

Crc16Dds110

Crc16DectR

Crc16DectX

Crc16Dnp

Crc16En13757

Crc16Genibus

Crc16Gsm

Crc16Ibm3740

Crc16IbmSdlc

Crc16IsoIec144433A

Crc16Kermit

Crc16Lj1200

Crc16M17

Crc16MaximDow

Crc16Mcrf4xx

Crc16Modbus

Crc16Nrsc5

Crc16OpensafetyA

Crc16OpensafetyB

Crc16Profibus

Crc16Riello

Crc16SpiFujitsu

Crc16T10Dif

Crc16Teledisk

Crc16Tms37157

Crc16Umts

Crc16Usb

Crc16Xmodem

Crc32Aixm

Crc32Autosar

Crc32Base91D

Crc32Bzip2

Crc32CdRomEdc

Crc32Cksum

Crc32Custom

👎Deprecated since 1.9.0:
Use CrcCustom instead, which works with any supported width (16, 32, 64)

Crc32Iscsi

Crc32IsoHdlc

Crc32Jamcrc

Crc32Mef

Crc32Mpeg2

Crc32Xfer

Crc64Custom

👎Deprecated since 1.9.0:
Use CrcCustom instead, which works with any supported width (16, 32, 64)

Crc64Ecma182

Crc64GoIso

Crc64Ms

Crc64Nvme

Crc64Redis

Crc64We

Crc64Xz