crc_fast

Function checksum_file_with_params

pub fn checksum_file_with_params(
    params: CrcParams,
    path: &str,
    chunk_size: Option<usize>,
) -> Result<u64, Error>

Computes the CRC checksum for the given file using custom CRC parameters.

Appears to be much faster (~2X) than using Writer and io::*, at least on Apple M2 Ultra

Errors

This function will return an error if the file cannot be read.

Examples

use std::env;
use crc_fast::{checksum_file_with_params, CrcParams};

// for example/test purposes only, use your own file path
let file_path = env::current_dir().expect("missing working dir").join("crc-check.txt");
let file_on_disk = file_path.to_str().unwrap();

// Define custom CRC-32 parameters (equivalent to CRC-32/ISO-HDLC)
let custom_params = CrcParams::new(
    "CRC-32/CUSTOM",
    32,
    0x04c11db7,
    0xffffffff,
    true,
    0xffffffff,
    0xcbf43926,
);

let checksum = checksum_file_with_params(custom_params, file_on_disk, None);

assert_eq!(checksum.unwrap(), 0xcbf43926);