crc_fast

Function checksum_file

pub fn checksum_file(
    algorithm: CrcAlgorithm,
    path: &str,
    chunk_size: Option<usize>,
) -> Result<u64, Error>

Computes the CRC checksum for the given file using the specified algorithm.

Appears to be much faster (~2X) than using Writer and io::*, at least on Apple M2 Ultra

Errors

This function will return an error if the file cannot be read.

Examples

checksum_file

 use std::env;
 use crc_fast::{checksum_file, CrcAlgorithm::Crc32IsoHdlc};

 // for example/test purposes only, use your own file path
 let file_path = env::current_dir().expect("missing working dir").join("crc-check.txt");
 let file_on_disk = file_path.to_str().unwrap();

 let checksum = checksum_file(Crc32IsoHdlc, file_on_disk, None);

 assert_eq!(checksum.unwrap(), 0xcbf43926);