crc_fast

Struct Digest

pub struct Digest { /* private fields */ }

Represents a CRC Digest, which is used to compute CRC checksums.

The Digest struct maintains the state of the CRC computation, including the current state, the amount of data processed, the CRC parameters, and the calculator function used to perform the CRC calculation.

Implementations

impl Digest

pub fn new(algorithm: CrcAlgorithm) -> Self

Creates a new Digest instance for the specified CRC algorithm.

Examples

use crc_fast::{Digest, CrcAlgorithm::Crc32IsoHdlc};

let mut digest = Digest::new(Crc32IsoHdlc);
digest.update(b"123456789");
let checksum = digest.finalize();

assert_eq!(checksum, 0xcbf43926);

pub fn new_with_init_state(algorithm: CrcAlgorithm, init_state: u64) -> Self

Creates a new Digest instance for the specified CRC algorithm with a custom initial state.

Examples

use crc_fast::{Digest, CrcAlgorithm::Crc32IsoHdlc};

// CRC-32/ISO-HDLC with initial state of 0x00000000, instead of the default initial state
// of 0xffffffff,
let mut digest = Digest::new_with_init_state(Crc32IsoHdlc, 0x00000000);
digest.update(b"123456789");
let checksum = digest.finalize();

// different initial state, so checksum will be different
assert_eq!(checksum, 0xd202d277);

let mut digest = Digest::new_with_init_state(Crc32IsoHdlc, 0xffffffff);
digest.update(b"123456789");
let checksum = digest.finalize();

// same initial state as the default, so checksum will be the same
assert_eq!(checksum, 0xcbf43926);

pub fn new_with_params(params: CrcParams) -> Self

Creates a new Digest instance with custom CRC parameters.

Examples

use crc_fast::{Digest, CrcParams};

// Define custom CRC-32 parameters (equivalent to CRC-32/ISO-HDLC)
let custom_params = CrcParams::new(
    "CRC-32/CUSTOM",
    32,
    0x04c11db7,
    0xffffffff,
    true,
    0xffffffff,
    0xcbf43926,
);

let mut digest = Digest::new_with_params(custom_params);
digest.update(b"123456789");
let checksum = digest.finalize();

assert_eq!(checksum, 0xcbf43926);

pub fn update(&mut self, data: &[u8])

Updates the CRC state with the given data.

pub fn finalize(&self) -> u64

Finalizes the CRC computation and returns the result.

pub fn finalize_reset(&mut self) -> u64

Finalizes the CRC computation, resets the state, and returns the result.

pub fn reset(&mut self)

Resets the CRC state to its initial value.

pub fn combine(&mut self, other: &Self)

Combines the CRC state with a second Digest instance.

pub fn get_amount(&self) -> u64

Gets the amount of data processed so far

pub fn get_state(&self) -> u64

Gets the current CRC state.

Examples

use crc_fast::{Digest, CrcAlgorithm::Crc32IsoHdlc};

let mut digest = Digest::new(Crc32IsoHdlc);
digest.update(b"123456789");
let state = digest.get_state();

// non-finalized state, so it won't match the final checksum
assert_eq!(state, 0x340bc6d9);

// finalized state will match the checksum
assert_eq!(digest.finalize(), 0xcbf43926);